Move printUserDetails() to print-users.ts
Refactor index.ts by moving the printUserDetails() into a new file print-users.ts & removing the optional chaining operator.

Rename index.test.ts to print-users.ts to match the new file.

Ran checkformat then format to ensure all files adhere to prettier guidelines.
