+++ .xo-config.json
@@ -2,7 +2,6 @@
   "extends": ["prettier"],
   "rules": {
     "no-useless-concat": "off",
-    "arrow-body-style": "off",
-    "no-unsafe-optional-chaining": "off"
+    "arrow-body-style": "off"
   }
 }
